requirements.txt
openai
requests
python-dotenv
beautifulsoup4
tldextract
markdownify
